fix: option cannot be selected when SVG triggers a click event (#45)
* fix: option cannot be selecting when SVG is present inside

We were checking if the target is an instance of `HTMLElement` and proceeding if it is. This will neglect elements such as SVGElement and others as they are not an instance of HTMLElement.
We simply need to check if the `target` exists. if it does, then proceed., else return `false`.

* update changelog